move ClinicalEntitySchemaNames to common model
