+++ src/clinical/donor-repo.ts
@@ -3,7 +3,7 @@ import mongoose from 'mongoose';
 import { DeepReadonly } from 'deep-freeze';
 import { F, MongooseUtils, notEmpty } from '../utils';
 import { loggerFor } from '../logger';
-import { ClinicalEntitySchemaNames } from '../submission/submission-entities';
+import { ClinicalEntitySchemaNames } from '../common-model/entities';
 
 export const SUBMITTER_ID = 'submitterId';
 export const SPECIMEN_SUBMITTER_ID = 'specimen.submitterId';
