+++ src/common-model/entities.ts
@@ -1,0 +1,11 @@
+export enum ClinicalEntitySchemaNames {
+  REGISTRATION = 'sample_registration',
+  DONOR = 'donor',
+  SPECIMEN = 'specimen',
+  PRIMARY_DIAGNOSIS = 'primary_diagnosis',
+  TREATMENT = 'treatment',
+  CHEMOTHERAPY = 'chemotherapy',
+  RADIATION = 'radiation',
+  FOLLOW_UP = 'follow_up',
+  HORMONE_THERAPY = 'hormone_therapy',
+}
