+++ src/schema/schema-api.ts
@@ -1,10 +1,10 @@
 import * as manager from './schema-manager';
 import { Request, Response } from 'express';
 import { loggerFor } from '../logger';
-import { SchemasDictionary, SchemaDefinition } from '../lectern-client/schema-entities';
+import { SchemaDefinition } from '../lectern-client/schema-entities';
 import { setStatus, Status } from '../app-health';
 import { ControllerUtils } from '../utils';
-import { ClinicalEntitySchemaNames } from '../submission/submission-entities';
+import { ClinicalEntitySchemaNames } from '../common-model/entities';
 import AdmZip from 'adm-zip';
 import { HasFullWriteAccess } from '../decorators';
 const L = loggerFor(__filename);
