+++ src/submission/submission-to-clinical/stat-calculator.ts
@@ -3,7 +3,7 @@ import {
   CoreCompletionStats,
   CoreClinicalEntites,
 } from '../../../src/clinical/clinical-entities';
-import { ClinicalEntitySchemaNames } from '../submission-entities';
+import { ClinicalEntitySchemaNames } from '../../common-model/entities';
 import { notEmpty, F } from '../../../src/utils';
 import { getClinicalEntitiesFromDonorBySchemaName } from './submission-to-clinical';
