+++ src/submission/validation-clinical/specimen.ts
@@ -2,13 +2,13 @@ import {
   SubmissionValidationError,
   DataValidationErrors,
   SubmittedClinicalRecord,
-  ClinicalEntitySchemaNames,
   DonorFieldsEnum,
   SpecimenFieldsEnum,
   ClinicalUniqueIdentifier,
   DonorVitalStatusValues,
   SampleRegistrationFieldsEnum,
 } from '../submission-entities';
+import { ClinicalEntitySchemaNames } from '../../common-model/entities';
 import { DeepReadonly } from 'deep-freeze';
 import { Donor, Specimen } from '../../clinical/clinical-entities';
 import * as utils from './utils';
