+++ src/submission/validation-clinical/therapy.ts
@@ -1,12 +1,11 @@
 import {
   SubmissionValidationError,
-  ClinicalEntitySchemaNames,
   TreatmentFieldsEnum,
   SubmittedClinicalRecord,
   DataValidationErrors,
   ClinicalTherapyType,
-  TherapyRxNormFields,
 } from '../submission-entities';
+import { ClinicalEntitySchemaNames } from '../../common-model/entities';
 import { DeepReadonly } from 'deep-freeze';
 import { Donor, Treatment } from '../../clinical/clinical-entities';
 import * as utils from './utils';
