+++ test/integration/submission/migration_utils/schema_builder.ts
@@ -7,7 +7,7 @@ import {
   ValueType,
 } from '../../../../src/lectern-client/schema-entities';
 import legacyStubSchemas from '../stub-schema.json';
-import { ClinicalEntitySchemaNames } from '../../../../src/submission/submission-entities';
+import { ClinicalEntitySchemaNames } from '../../../../src/common-model/entities';
 import * as fieldNames from './fields';
 
 const DICTIONARY_NAME = 'ARGO Clinical Submission';
